limit to long change

+++ ApiRateLimiterUserIdAltairCA/Interface/IAPIRateLimiterUserIdHttpFilterService.cs
@@ -7,7 +7,7 @@ namespace APIRateLimiterUserId.AspNetCore.AltairCA.Interface
 {
     internal interface IAPIRateLimiterUserIdHttpFilterService
     {
-        Task<Tuple<bool, APIRateLimiterUserIdServiceResponse>> Validate(TimeSpan span, int limit, string groupKey);
+        Task<Tuple<bool, APIRateLimiterUserIdServiceResponse>> Validate(TimeSpan span, long limit, string groupKey);
         Task<Tuple<bool, APIRateLimiterUserIdServiceResponse>> Validate();
         void SetHeaderAndBodyIfLimitReached(ActionExecutingContext context, APIRateLimiterUserIdServiceResponse response);
         void SetHeadersIfNotLimitReached(ActionExecutedContext context, APIRateLimiterUserIdServiceResponse response);
