+++ ApiRateLimiterUserIdAltairCA/Models/APIRateLimiterUserIdServiceResponse.cs
@@ -4,9 +4,9 @@ namespace APIRateLimiterUserId.AspNetCore.AltairCA.Models
 {
     public class APIRateLimiterUserIdServiceResponse
     {
-        public int AvaliableLimit { get; set; }
+        public long AvaliableLimit { get; set; }
         public DateTime ResetIn { get; set; }
-        public int MaxLimit { get; set; }
+        public long MaxLimit { get; set; }
         public double Period { get; set; }
     }
 }
